# atlas/contracts/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Atlas API
  version: 1.0.0
  description: |
    Contrato HTTP da plataforma Atlas.

    Este arquivo é a FONTE DE VERDADE do transporte. Ele deve permanecer em
    paridade com os ports em `packages/api-client/src/ports/` — o CI compara os
    paths declarados aqui com os métodos dos ports. Divergência é erro de build.

    Regra de evolução (spec 30 §6): adicionar campo opcional ou endpoint é
    permitido; remover, renomear ou estreitar tipo exige /api/v2 e 90 dias de
    convivência. Apps móveis não atualizam sincronizadamente.
servers:
  - url: https://api.atlas.app
    description: Produção
  - url: http://localhost:5080
    description: Desenvolvimento
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/exercises:
    get:
      tags:
        - Catalog
      operationId: listExercises
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: muscle
          in: query
          schema:
            type: string
        - name: equipment
          in: query
          schema:
            $ref: "#/components/schemas/Equipment"
        - name: difficulty
          in: query
          schema:
            $ref: "#/components/schemas/Difficulty"
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - nextCursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ExerciseSummary"
                  nextCursor:
                    type:
                      - string
                      - "null"
  /api/v1/exercises/{id}:
    get:
      tags:
        - Catalog
      operationId: getExercise
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseDetail"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    NotFound:
      description: Recurso não encontrado
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    Problem:
      type: object
      description: RFC 9457
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
...
    Equipment:
      type: string
      enum:
        - barbell
        - dumbbell
        - machine
        - cable
        - bodyweight
        - kettlebell
        - band
        - ball
        - none
    Difficulty:
      type: string
      enum:
        - beginner
        - intermediate
        - advanced
    MuscleRole:
      type: string
      enum:
        - primary
        - secondary
        - stabilizer
...
    MuscleActivation:
      type: object
      required:
        - muscleGroupId
        - muscleCode
        - role
        - activationWeight
      properties:
        muscleGroupId:
          type: integer
        muscleCode:
          type: string
        role:
          $ref: "#/components/schemas/MuscleRole"
        activationWeight:
          type: number
          minimum: 0
          maximum: 1
          description: Multiplicador do volume ponderado que alimenta o heatmap.
    ExerciseSummary:
      type: object
      required:
        - id
        - name
        - primaryMuscleCode
        - equipment
        - difficulty
        - thumbnailUrl
        - isCustom
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        primaryMuscleCode:
          type: string
        equipment:
          $ref: "#/components/schemas/Equipment"
        difficulty:
          $ref: "#/components/schemas/Difficulty"
        thumbnailUrl:
          type:
            - string
            - "null"
        isCustom:
          type: boolean
    ExerciseDetail:
      allOf:
        - $ref: "#/components/schemas/ExerciseSummary"
        - type: object
          required:
            - description
            - media
            - activations
            - executionCues
            - commonMistakes
          properties:
            description:
              type: string
            media:
              type: array
              items:
                type: object
                required:
                  - kind
                  - url
                properties:
                  kind:
                    type: string
                    enum:
                      - loop
                      - video
                      - image
                  url:
                    type: string
                    format: uri
                  angle:
                    type: string
                    enum:
                      - front
                      - side
                      - rear
                  durationMs:
                    type: integer
            activations:
              type: array
              minItems: 1
              items:
                $ref: "#/components/schemas/MuscleActivation"
            executionCues:
              type: array
              items:
                type: string
            commonMistakes:
              type: array
              items:
                type: string
...